Use viewmodel for file picker.
