+++ Scrawler/View/PageOptionsDialog.xaml.cs
@@ -19,33 +19,22 @@ namespace Scrawler.View
             this.InitializeComponent();
         }
 
-        private void ContentDialog_OkButtonClick(ContentDialog sender, 
+        private void ContentDialog_OkButtonClick(ContentDialog sender,
             ContentDialogButtonClickEventArgs args)
         {
         }
 
-        private void ContentDialog_CancelButtonClick(ContentDialog sender, 
+        private void ContentDialog_CancelButtonClick(ContentDialog sender,
             ContentDialogButtonClickEventArgs args)
         {
         }
 
-        private async void OpenImageButton_Click(object sender, Windows.UI.Xaml.RoutedEventArgs e)
+        private void OpenImageButton_Click(object sender, Windows.UI.Xaml.RoutedEventArgs e)
         {
             var viewModel = DataContext as PageOptionsViewModel;
-            if (viewModel == null) return;
-
-            var picker = new FileOpenPicker();
-            picker.FileTypeFilter.Add(".png");
-
-            var file = await picker.PickSingleFileAsync();
-            if (file == null) return;
-
-            // Ensure the stream is disposed once the image is loaded
-            using (var fileStream = await file.OpenAsync(Windows.Storage.FileAccessMode.Read))
+            if (viewModel != null)
             {
-                var device = CanvasDevice.GetSharedDevice();
-                var bitmapImage = await CanvasBitmap.LoadAsync(device, fileStream);
-                viewModel.LoadImageForBackground(bitmapImage);
+                viewModel.LoadImageForBackground();
             }
         }
     }
